chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-dotenv
requests
transformers
torch
sentence-transformers
faiss-cpu
python-multipart
gpt4all

